tests: split validating webhook integration tests to a separate CI job
