+++ test/integration/validatingwebhook/suite_test.go
@@ -1,0 +1,11 @@
+package validatingwebhook
+
+import (
+	"testing"
+
+	"github.com/kong/kong-operator/test/integration"
+)
+
+func TestMain(m *testing.M) {
+	integration.Suite(m)
+}
