+++ test/integration/validatingwebhook/validating_secretcredentials_test.go
@@ -1,4 +1,4 @@
-package integration
+package validatingwebhook
 
 import (
 	"context"
@@ -20,13 +20,13 @@ import (
 	"github.com/kong/kong-operator/controller/konnect"
 	"github.com/kong/kong-operator/internal/annotations"
 	"github.com/kong/kong-operator/modules/manager/config"
+	"github.com/kong/kong-operator/test/integration"
 )
 
 func TestAdmissionWebhook_SecretCredentials(t *testing.T) {
-	t.Parallel()
-
+	ctx := t.Context()
 	_, cleaner, ingressClass, ctrlClient := bootstrapGateway(
-		t.Context(), t, env, GetClients().MgrClient,
+		ctx, t, integration.GetEnv(), integration.GetClients().MgrClient,
 	)
 
 	// highEndConsumerUsageCount indicates a number of consumers with credentials
